// src/services/listed-bond/dto/get-product.dto.ts
export type GetProductResp = {
    ISIN: string;
    aum: string;
    baseXirr: string;
    bondId: string;
    category: string;
    comms: {
        waMediaURL: string;
    };
    couponRate: number;
    couponType: string;
    descLogoUrl: string;
    description: string;
    digio: {
        sellerSignatureText: string;
        sellerName: string;
        sellerDpName: string;
        sellerDpID: string;
        sellerClientId: string;
        securityDetails: string;
        modeOfIssue: string;
        message: string;
        issuer: string;
        expiryInDays: number;
        depository: string;
        sellerDepository: string;
        templateId: string;
        cin: string;
        debentureTrustee: string;
    };
    displayName: string;
    exitLoadInPerc: number;
    faceValue: number;
    fundId: string;
    gnpa: string;
    highlights: {
        web: string;
        text: string;
    }[];
    isActive: boolean;
    issueCategory: string;
    issueDate: string;
    issuePrice: number;
    issueSize: number;
    issueType: string;
    issuer: string;
    issuerFaceValue: number;
    lenderIcons: string[];
    listing: string;
    listingCategory: string;
    listingType: string;
    logoUrl: string;
    maturityDate: string;
    maxAmt: number;
    maxReturns: number;
    maxTenure: number;
    maxXirrDeviation: string;
    minAmt: number;
    minReturns: number;
    minTenure: number;
    minUnits: number;
    minXirrDeviation: string;
    multiplier: number;
    netWorth: string;
    openDate: string;
    order: number;
    priceIsLive: boolean;
    principalRepayment: string;
    product: string;
    productName: string;
    productType: string;
    profitAfterTax: string;
    rating: string;
    saleTag: null;
    securityCover: string;
    soldOutTxt: string;
    stepSize: number;
    tagLine: string;
    taxSaving: string;
    tds: number;
    trueISIN: string;
    whyInvest: string;
    xirr: string;
    xirrPriceForUser: boolean;
    closeDate: string;
    companyFinancialsLink: string;
    companyPresentationLink: string;
    debtEquityRatio: string;
    marketCap: string;
    marketEndTime: string;
    marketStartTime: string;
    marketingGifBanner: string;
    nnpa: string;
    productInfoMemorandum: string;
    prospectusLink: string;
    ratingRationaleUrl: string;
    showOnBrowse: boolean;
    isTxnAllowed: boolean;
    showOnMobile: boolean;
};